initial pass at creating a template content space entity
