+++ src/domain/template/template-content-space/dto/template.content.space.dto.create.ts
