--- src/domain/template/template-content-space/index.ts
@@ -1,5 +1,0 @@
-export * from './dto/templateContentSpace.dto.update';
-export * from './dto/templateContentSpace.dto.delete';
-export * from './dto/templateContentSpace.dto.create';
-export * from '../templateContentSpace.defaults/definitions/templateContentSpace.community.roles';
-export * from '../templateContentSpace.defaults/definitions/templateContentSpace.community.role.application.form';
